Move Demos to separate file
